refactor: streamline wallet event subscription and model integration
